requirements.txt
aiofiles
aiohttp
aiosignal
async-timeout
attrs
beautifulsoup4
biliass
blinker
Brotli
certifi
charset-normalizer
click
colorama
dicttoxml
filelock
Flask
frozenlist
greenlet
idna
importlib-metadata
itsdangerous
Jinja2
lxml
MarkupSafe
multidict
mutagen
platformdirs
playwright
protobuf
pycryptodomex
pyee
PyYAML
redis
requests
soupsieve
typing_extensions
uritools
urlextract
urllib3
websockets
Werkzeug
wget
yarl
yt-dlp
zipp
